metal-app: Derive a camera's projection matrix based a fixed 60 degree vertical FOV.
This fixes a window resizing issue. Previously, as you widen the screen, everything is scaled up, which feels unnatural/non-standard (quite pronounced with proj-6 environment mapped background). This was due to the calculations were based on fixed-ish Horizontal FOV and changing the Vertical FOV based on that and the aspect ratio.

Another issue was the Horizontal FOV was dynamicaly calculated based on model's bounding box. This was slapped in to make sure any model loaded would be mostly visible (only considered the model's width). Turns out to be unnecessary as we scaled the model to a [0,1] range right off the bat (to reduce precision errors, revealed in Yoda model).